#2 refactor: JWT 파일 구조 변경
- Global JWT 관련 파일 구조 변경
- authentication -> jwt
- infra -> provider
