+++ src/main/java/com/hyundai/softeer/backend/domain/user/dto/LoginResponseDto.java
@@ -1,6 +1,6 @@
 package com.hyundai.softeer.backend.domain.user.dto;
 
-import com.hyundai.softeer.backend.global.authentication.domain.TokenDto;
+import com.hyundai.softeer.backend.global.jwt.dto.TokenDto;
 import lombok.Builder;
 import lombok.Data;
