--- src/main/java/com/hyundai/softeer/backend/global/authentication/domain/oauth/OAuthProvider.java
@@ -1,5 +1,0 @@
-package com.hyundai.softeer.backend.global.authentication.domain.oauth;
-
-public enum OAuthProvider {
-    KAKAO, NAVER
-}
