+++ src/main/java/com/hyundai/softeer/backend/global/jwt/OAuthApiClient.java
@@ -1,4 +1,4 @@
-package com.hyundai.softeer.backend.global.authentication.domain.oauth;
+package com.hyundai.softeer.backend.global.jwt;
 
 public interface OAuthApiClient {
     OAuthProvider oAuthProvider();
