+++ src/main/java/com/hyundai/softeer/backend/global/jwt/OAuthProvider.java
@@ -1,0 +1,5 @@
+package com.hyundai.softeer.backend.global.jwt;
+
+public enum OAuthProvider {
+    KAKAO, NAVER
+}
