+++ src/main/java/com/hyundai/softeer/backend/global/jwt/provider/kakao/KakaoTokens.java
@@ -1,4 +1,4 @@
-package com.hyundai.softeer.backend.global.authentication.infra.naver;
+package com.hyundai.softeer.backend.global.jwt.provider.kakao;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -6,17 +6,23 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class NaverTokens {
+public class KakaoTokens {
 
     @JsonProperty("access_token")
     private String accessToken;
 
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-
     @JsonProperty("token_type")
     private String tokenType;
 
+    @JsonProperty("refresh_token")
+    private String refreshToken;
+
     @JsonProperty("expires_in")
     private String expiresIn;
+
+    @JsonProperty("refresh_token_expires_in")
+    private String refreshTokenExpiresIn;
+
+    @JsonProperty("scope")
+    private String scope;
 }
