+++ src/main/java/com/hyundai/softeer/backend/global/jwt/provider/naver/NaverTokens.java
@@ -1,4 +1,4 @@
-package com.hyundai.softeer.backend.global.authentication.infra.kakao;
+package com.hyundai.softeer.backend.global.jwt.provider.naver;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -6,23 +6,17 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class KakaoTokens {
+public class NaverTokens {
 
     @JsonProperty("access_token")
     private String accessToken;
 
-    @JsonProperty("token_type")
-    private String tokenType;
-
     @JsonProperty("refresh_token")
     private String refreshToken;
 
+    @JsonProperty("token_type")
+    private String tokenType;
+
     @JsonProperty("expires_in")
     private String expiresIn;
-
-    @JsonProperty("refresh_token_expires_in")
-    private String refreshTokenExpiresIn;
-
-    @JsonProperty("scope")
-    private String scope;
 }
